MAINT: Rename internal _openmeeg_cxx -> _openmeeg_wrapper
