+++ .github/workflows/build_and_test.yml
@@ -483,7 +483,7 @@ jobs:
             export PYTHONPATH=$PWD/install/lib/${PYNAME}${PYVER}/site-packages
           fi
           echo "PYTHONPATH=$PYTHONPATH"
-          python -c "import openmeeg._openmeeg_cxx as _omc; print(_omc.sqr(4)); assert _omc.sqr(4) == 16"
+          python -c "import openmeeg._openmeeg_wrapper as _omc; print(_omc.sqr(4)); assert _omc.sqr(4) == 16"
         fi
 
     - name: Create, delocate, check, install, and test wheel
