+++ wrapping/python/setup.cfg
@@ -1,6 +1,6 @@
 [flake8]
 ignore = W503
-exclude = openmeeg/_openmeeg_cxx.py
+exclude = openmeeg/_openmeeg_wrapper.py
 per-file-ignores =
     __init__.py:F401
     setup.py:E501
